backend/requirements.txt
FastAPI
firebase-admin
openai
uvicorn
python-dotenv
itsdangerous
flake8
isort
ruff